Refactor view file structure
